Move geo macros to a separate module
